Add Exchange to Deployment
This patch removes the info of other parties from the proxy's route
table. Instead, an `exchange` is added to maintain the entry point of
all parties, so that any party could find out another party through
exchange easily.

1. update `docker-auto-deploy.sh` to handle the deployment of exchange. By
default, the exchange will be deployed on the first party defined in
`partylist`.

2. update fate_flow's properties

Signed-off-by: jiahaoc <jiahaoc@vmware.com>