Signed-off-by: Chenyang Gao <gps949@outlook.com> (#670)
Add doc about how to config TLS kubefate service

Few days ago, I made a PR to remove the hardcode URL scheme in the cli.
To give more help to the guys want to deploy the kubefate service with a TLS, I add a markdown doc as tutorial.
It's not so detail. But may be a very first step for the ones may be interested in.